refactored idex buffer

+++ market_economy_simulation/src/geometry/circle.rs
@@ -12,7 +12,7 @@ use crate::deferred_color_shader;
 pub struct Circle {
     pub vertices: Vec<Vertex>,
     pub colors: Vec<Color>,
-    pub indices: Vec<u32>,
+    pub indices: Vec<u16>,
 
     pub deferred_vertices: Vec<deferred_color_shader::Vertex>,
 }
@@ -22,7 +22,7 @@ impl Circle {
     pub fn new(r: f32, n: usize) -> Self {
         let mut vertices = Vec::<Vertex>::new();
         let mut colors = Vec::<Color>::new();
-        let mut indices = Vec::<u32>::new();
+        let mut indices = Vec::<u16>::new();
 
         let mut deferred_vertices = Vec::<deferred_color_shader::Vertex>::new();
 
@@ -52,13 +52,13 @@ impl Circle {
 
         for i in 1..n {
             indices.push(0);
-            indices.push(i as u32);
-            indices.push((i + 1) as u32);
+            indices.push(i as u16);
+            indices.push((i + 1) as u16);
         }
 
         indices.push(0);
-        indices.push((n) as u32);
-        indices.push(1_u32);
+        indices.push((n) as u16);
+        indices.push(1_u16);
 
         let normal = [0.0, 0.0, 1.0];
         for vertex in &vertices {
